Menyelesaikan Fitur Penawaran dan Penjualan

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use App\Models\MediaCategory;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -17,9 +18,10 @@ class UserController extends Controller
     public function index(): Response
     {
         if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Owner' ){
-            return response()->view('dashboard.users.users.index', [
-                'users' => User::sortable()->paginate(10),
-                'title' => 'Daftar User'
+            return response()->view('users.index', [
+                'users' => User::filter(request('search'))->sortable()->paginate(10)->withQueryString(),
+                'title' => 'Daftar User',
+                'categories' => MediaCategory::all()
             ]);
         } else {
             abort(403);
@@ -33,8 +35,9 @@ class UserController extends Controller
     {
         $this->authorize('isAdmin');
 
-        return response()->view('dashboard.users.users.create', [
-            'title' => 'Tambah User'
+        return response()->view('users.create', [
+            'title' => 'Menambah Data Pengguna',
+            'categories' => MediaCategory::all()
         ]);
     }
 
@@ -44,16 +47,23 @@ class UserController extends Controller
     public function store(Request $request): RedirectResponse
     {
         $this->authorize('isAdmin');
+        if($request->password != $request->confirm_password){
+            return back()->withErrors(['confirm_password' => ['Konfirmasi password tidak sesuai']])->withInput();
+        }
 
-        if ($request->gender == 'Pilih Jenis Kelamin'){
+        if ($request->gender == 'pilih'){
             return back()->withErrors(['gender' => ['Silahkan pilih jenis kelamin']])->withInput();
         }
-        if ($request->level == 'Pilih Divisi'){
-            return back()->withErrors(['level' => ['Silahkan pilih divisi']])->withInput();
+        if ($request->division == 'pilih'){
+            return back()->withErrors(['division' => ['Silahkan pilih divisi']])->withInput();
         }
-        if ($request->position == 'Pilih Jabatan'){
+        if ($request->position == 'pilih'){
             return back()->withErrors(['position' => ['Silahkan pilih jabatan']])->withInput();
         }
+        if ($request->level == 'pilih'){
+            return back()->withErrors(['level' => ['Silahkan pilih level akses']])->withInput();
+        }
+
         $validateData = $request->validate([
             'name' => 'required|max:255',
             'username' => 'required|min:6|unique:users',
@@ -61,6 +71,7 @@ class UserController extends Controller
             'phone' => 'required|unique:users',
             'gender' => 'required',
             'level' => 'required',
+            'division' => 'required',
             'position' => 'required',
             'password' => 'required|min:8',
             'avatar' => 'image|file|max:1024'
@@ -76,7 +87,7 @@ class UserController extends Controller
 
         User::create($validateData);
         
-        return redirect('/dashboard/users/users')->with('success','User baru '. $request->username . ' berhasil ditambahkan');
+        return redirect('/users')->with('success','Data pengguna baru dengan nama'. $request->name . ' berhasil ditambahkan');
         
     }
 
@@ -86,14 +97,16 @@ class UserController extends Controller
     public function show(User $user): Response
     {
         if($user->id === auth()->user()->id){
-            return response()->view('dashboard.users.users.show', [
+            return response()->view('users.show', [
                 'user' => $user,
-                'title' => 'Detail User'
+                'title' => 'Detail User',
+                'categories' => MediaCategory::all()
             ]);
         } else if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Owner' ){
-            return response()->view('dashboard.users.users.show', [
+            return response()->view('users.show', [
                 'user' => $user,
-                'title' => 'Detail User'
+                'title' => 'Detail User',
+                'categories' => MediaCategory::all()
             ]);
         } else {
             abort(401);
@@ -105,18 +118,20 @@ class UserController extends Controller
     public function edit(User $user): Response
     {
         if($user->id === auth()->user()->id){
-            return response()->view('dashboard.users.users.edit', [
+            return response()->view('users.edit', [
                 'user' => $user,
-                'title' => 'Edit User'
+                'title' => 'Edit User',
+                'categories' => MediaCategory::all()
             ]);
         } else {
-        $this->authorize('isAdmin');
-        
-        return response()->view('dashboard.users.users.edit', [
-            'user' => $user,
-            'title' => 'Edit User'
-        ]);
-    }
+            $this->authorize('isAdmin');
+            
+            return response()->view('users.edit', [
+                'user' => $user,
+                'title' => 'Edit User',
+                'categories' => MediaCategory::all()
+            ]);
+        }
     }
 
     /**
@@ -126,6 +141,10 @@ class UserController extends Controller
     {
         $rules = [
             'name' => 'required|max:255',
+            'gender' => 'required',
+            'level' => 'required',
+            'division' => 'required',
+            'position' => 'required',
             'avatar' => 'image|file|max:1024'
         ];
 
@@ -140,23 +159,8 @@ class UserController extends Controller
         if($request->phone != $user->phone){
             $rules['phone'] = 'required|unique:users';
         }
-
-        if($request->position != $user->position){
-            $rules['position'] = 'required';
-        }
-
-        if($request->level != $user->level){
-            $rules['level'] = 'required';
-        }
-
-        if($request->gender != $user->gender){
-            $rules['gender'] = 'required';
-        }
-
         if($request->password){
             $rules['password'] = 'required|min:8';
-        } else {
-            $rules['password'] = $request->oldPassword;
         }
 
         $validateData = $request->validate($rules);
@@ -176,9 +180,9 @@ class UserController extends Controller
                 ->update($validateData);
 
         if(auth()->user()->level === 'Administrator'){
-            return redirect('/dashboard/users/users')->with('success','User Has Been Updated');
+            return redirect('/users')->with('success','Data pengguna dengan nama ' . $user->name. ' berhasil diupdate');
         } else {
-            return redirect('/dashboard/users/users/' . $user->id)->with('success','User Has Been Updated');
+            return redirect('/users/' . $user->id)->with('success','Data pengguna dengan nama ' . $user->name. ' berhasil diupdate');
         }
     }
 
@@ -188,6 +192,11 @@ class UserController extends Controller
     public function destroy(User $user): RedirectResponse
     {
         $this->authorize('isAdmin');
+
+        $dataUsers = User::where('level', 'Administrator')->get();
+        if(count($dataUsers) == 1){
+            return back()->withErrors(['delete' => ['Gagal untuk menghapus data pengguna, minimal harus ada 1 pengguna dengan level Administrator']]);
+        }
         
         if($user->avatar){
             Storage::delete($user->avatar);
@@ -195,6 +204,6 @@ class UserController extends Controller
 
         User::destroy($user->id);
 
-        return redirect('/dashboard/users/users')->with('success','User ' . $user->username . ' berhasil dihapus');
+        return redirect('/users')->with('success','Data penggunan dengan nama ' . $user->name . ' berhasil dihapus');
     }
 }
